+++ src/snowflake/snowpark/modin/plugin/extensions/index.py
@@ -398,7 +398,6 @@ class Index(metaclass=TelemetryMeta):
         return self.to_pandas().values
 
     @property
-    @index_not_implemented()
     def is_monotonic_increasing(self) -> None:
         """
         Return a boolean if the values are equal or increasing.
@@ -411,11 +410,23 @@ class Index(metaclass=TelemetryMeta):
         See Also
         --------
         Index.is_monotonic_decreasing : Check if the values are equal or decreasing
+
+        Examples
+        --------
+        >>> pd.Index([1, 2, 3]).is_monotonic_increasing
+        True
+        >>> pd.Index([1, 2, 2]).is_monotonic_increasing
+        True
+        >>> pd.Index([1, 3, 2]).is_monotonic_increasing
+        False
         """
-        # TODO: SNOW-1458134 implement is_monotonic_increasing
+        return (
+            self._query_compiler.is_monotonic_increasing(index=True)
+            .to_pandas()
+            .squeeze()
+        )
 
     @property
-    @index_not_implemented()
     def is_monotonic_decreasing(self) -> None:
         """
         Return a boolean if the values are equal or decreasing.
@@ -428,8 +439,21 @@ class Index(metaclass=TelemetryMeta):
         See Also
         --------
         Index.is_monotonic_increasing : Check if the values are equal or increasing
+
+        Examples
+        --------
+        >>> pd.Index([3, 2, 1]).is_monotonic_decreasing
+        True
+        >>> pd.Index([3, 2, 2]).is_monotonic_decreasing
+        True
+        >>> pd.Index([3, 1, 2]).is_monotonic_decreasing
+        False
         """
-        # TODO: SNOW-1458134 implement is_monotonic_decreasing
+        return (
+            self._query_compiler.is_monotonic_decreasing(index=True)
+            .to_pandas()
+            .squeeze()
+        )
 
     @property
     def is_unique(self) -> bool: